pyproject.toml
[project]
name = "damaage-panel-recog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "eclipse-zenoh==1.6.2",
  "motpy>=0.0.10",
  "msg",
  "numpy>=2.4.2",
  "opencv-python>=4.13.0.92",
  "protobuf>=6.33.5",
  "protovalidate>=1.1.1",
  "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "types-protobuf>=6.32.1.20251210",
]

[tool.uv.workspace]
members = [
    "msg",
]

[tool.uv.sources]
msg = { workspace = true }
